[Refactor] Integration
- 제니의 퍼블리싱 작업을 합침
- 불필요한 주석 제거
