+++ banchan/package.json
@@ -13,6 +13,7 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-scripts": "4.0.3",
+    "reallyawesome-jennyrousel": "^0.1.0",
     "styled-components": "^5.2.3",
     "web-vitals": "^1.0.1"
   },
