+++ banchan/src/components/utils/modal/ModalCard.jsx
@@ -1,8 +1,8 @@
 import React, { useState } from "react";
 import styled from "styled-components";
-import Price from "../Price";
-import { LabelList, StyledDescription, StyledTitle } from "../styles/common";
-import TextButton from "../button/TextButton";
+// import Price from "../Price";
+// import { LabelList, StyledDescription, StyledTitle } from "../styles/common";
+// import TextButton from "../button/TextButton";
 import ProductImage from "./products/ProductImage";
 import ProductInformation from "./products/ProductInformation";
 
@@ -10,11 +10,16 @@ const ModalCard = ({ product }) => {
   const [count, setCount] = useState(1);
   const [productImg, setProductImg] = useState(product.top_image);
 
-  const thumbnailList = [...new Array(5)].map((i, index) =>
+  const thumbnailList = [...Array(5)].map((item, index) =>
     index >= product.thumb_images.length ? (
       <BlackThumbnail />
     ) : (
-      <Thumbnail src={product.thumb_images[index]} onClick={() => setProductImg(product.thumb_images[index])} />
+      <Thumbnail
+        src={product.thumb_images[index]}
+        onClick={() => {
+          setProductImg(product.thumb_images[index]);
+        }}
+      />
     )
   );
 
@@ -26,36 +31,8 @@ const ModalCard = ({ product }) => {
 
   return (
     <CardContent>
-      <ProductImage imageUrl={product.top_image} thumbnailList={thumbnailList} />
-      {/* <ProductImageSection>
-        <img src={product.top_image} alt="product-thumbnail" />
-        <ThumbnailUL>
-          {product.thumb_images.map((imageUrl) => (
-            <li>
-              <Thumbnail src={imageUrl} />
-            </li>
-          ))}
-        </ThumbnailUL>
-      </ProductImageSection> */}
+      <ProductImage imageUrl={productImg} thumbnailList={thumbnailList} />
       <ProductInformation {...{ product, getTotalPrice, increaseCount, decreaseCount, count }} />
-      {/* <Information>
-        <ProductMainInfo>
-          <StyledTitle>{product.title}</StyledTitle>
-          <StyledDescription>{product.description}</StyledDescription>
-          <div>
-            <LabelList />
-            <Price product={product} />
-          </div>
-        </ProductMainInfo>
-        <ProductBuyInfo>
-          <div>적립금: {product.point}</div>
-          <div>배송정보: {product.delivery_info}</div>
-          <div>배송비 : {product.delivery_fee}</div>
-        </ProductBuyInfo>
-        <ProductCount></ProductCount>
-        <ProductPrice>여기는 총 주문 금액이 들어갈 예정입니다. 카운트를 같이 계산해서..</ProductPrice>
-        <TextButton type="ORDER"></TextButton>
-      </Information> */}
     </CardContent>
   );
 };
@@ -64,17 +41,12 @@ export default ModalCard;
 
 const CardContent = styled.div`
   background: white;
-  width: 960px;
+  // width: 960px;
+  width: fit-content;
   height: 1076px;
   padding: 48px;
   display: flex;
 `;
-
-// const CardContent = styled.div`
-//   background: white;
-//   width: 960px;
-//   height: 1076px;
-// `;
 
 const BlackThumbnail = styled.div`
   width: 72px;
@@ -91,25 +63,3 @@ const Thumbnail = styled.img`
   margin: 8px 8px 8px 0;
   cursor: pointer;
 `;
-// const ThumbnailUL = styled.div`
-//   display: flex;
-// `;
-
-// const Thumbnail = styled.img`
-//   width: 100px;
-//   height: 100px;
-// `;
-
-// const ProductImageSection = styled.section`
-//   margin-right: 32px;
-// `;
-
-// const Information = styled.div``;
-
-// const ProductMainInfo = styled.div``;
-
-// const ProductBuyInfo = styled.div``;
-
-// const ProductCount = styled.div``;
-
-// const ProductPrice = styled.div``;
